requirements.txt
apache-airflow==2.8.1
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
argparse==1.4.0
 
